refactor: LearnArticlePage with slug routing and article registry
Adds /learn/:slug as a dedicated article view. Slug → component is
resolved through articleRegistry; missing slugs redirect back to the
index. The article shell adds a Learn breadcrumb, prev/next links to
the next published or draft article in the same section, and a Draft
banner for in-review drafts.

Status vocabulary moves from live/stub to published/draft/soon.
Strokes-gained, benchmarks, and glossary content are extracted out of
the old UnderstandingTheGame section and into article components so
the registry can route to them. The legacy section components, the
in-page table of contents, and the active-section scroll spy were
only used by the long-scroll layout and are removed.
