+++ apps/web/src/pages/learn/data/learnSections.ts
@@ -1,4 +1,4 @@
-export type ArticleStatus = 'live' | 'stub'
+export type ArticleStatus = 'published' | 'draft' | 'soon'
 
 export interface ArticleStub {
   id: string
@@ -27,7 +27,7 @@ export const LEARN_SECTIONS: LearnSection[] = [
         title: 'How strokes gained works',
         description:
           'How every shot is graded against an expectation, and what +0.3 SG-Putting actually means.',
-        status: 'live',
+        status: 'published',
         words: 720,
       },
       {
@@ -35,7 +35,7 @@ export const LEARN_SECTIONS: LearnSection[] = [
         title: 'Reading your stats',
         description:
           'Where each stat lands across the field, from a 25-handicap weekend round up to the PGA Tour.',
-        status: 'live',
+        status: 'published',
         words: 200,
       },
       {
@@ -43,7 +43,7 @@ export const LEARN_SECTIONS: LearnSection[] = [
         title: 'Glossary of golf terms',
         description:
           'Handicap, GIR, scrambling, up-and-down, sand save, dispersion — the working vocabulary of the app.',
-        status: 'live',
+        status: 'published',
         words: 700,
       },
     ],
@@ -57,19 +57,19 @@ export const LEARN_SECTIONS: LearnSection[] = [
         id: 'guide-to-fittings',
         title: 'Guide to golf fittings',
         description: 'What a fitting actually changes and when one is worth the money.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'training-aids',
         title: 'Training aids explained',
         description: 'Which devices teach a real skill and which ones just feel productive.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'building-your-bag',
         title: 'Building your bag',
         description: 'Picking 14 clubs that cover your distances without dead zones.',
-        status: 'stub',
+        status: 'soon',
       },
     ],
   },
@@ -83,44 +83,44 @@ export const LEARN_SECTIONS: LearnSection[] = [
         title: 'How to practice effectively',
         description:
           'Block, random, variable, pressure — what each kind of practice is for and when to use it.',
-        status: 'live',
+        status: 'draft',
         words: 2700,
       },
       {
         id: 'practice-modes',
         title: 'Block, random, and pressure practice',
         description: 'A deeper look at the three modes and how to combine them.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'measurable-goals',
         title: 'Creating measurable practice goals',
         description: 'Turning "work on my irons" into something you can actually pass or fail.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'skill-and-pressure-games',
         title: 'Skill games and pressure games',
         description: 'Range games that build skills that survive the first tee.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'understanding-your-swing',
         title: 'Understanding your own swing',
         description: 'Reading your own miss patterns without an instructor in the bay.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'swing-variations',
         title: 'Swing variations for different body types',
         description: 'Why the textbook swing fits some bodies and not others.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'operation-36',
         title: 'The Operation 36 philosophy',
         description: 'Building a game from inside out, par at every distance.',
-        status: 'stub',
+        status: 'soon',
       },
     ],
   },
@@ -134,26 +134,26 @@ export const LEARN_SECTIONS: LearnSection[] = [
         title: 'Course management guide',
         description:
           'Way of the Playa, the Scoring Zone, planning a hole backwards. The mental side of shooting your number.',
-        status: 'live',
+        status: 'draft',
         words: 2200,
       },
       {
         id: 'mental-game',
         title: 'Mental game and on-course psychology',
         description: 'Routines, target focus, recovering from a bad shot mid-round.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'practice-vs-scoring-round',
         title: 'Practice round vs scoring round',
         description: 'Two different activities you should never mix.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'self-diagnosis',
         title: 'Self-diagnosis: finding your weaknesses',
         description: 'Using your own SG and pattern data to know where strokes are leaking.',
-        status: 'stub',
+        status: 'soon',
       },
     ],
   },
@@ -166,19 +166,19 @@ export const LEARN_SECTIONS: LearnSection[] = [
         id: 'lessons-and-coaching',
         title: 'Guide to lessons and coaching',
         description: 'Picking an instructor and getting a return on the lesson hour.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'fittings-with-coaches',
         title: 'Guide to golf fittings (all types)',
         description: 'Driver, iron, wedge, putter — what each fitting covers.',
-        status: 'stub',
+        status: 'soon',
       },
       {
         id: 'questions-for-coach',
         title: 'Questions to ask your coach',
         description: 'The conversation that makes a lesson stick beyond the bay.',
-        status: 'stub',
+        status: 'soon',
       },
     ],
   },
